This is a merge of master 01/07

+++ host/linux/time.c
@@ -3,6 +3,7 @@
 
 #include <errno.h>
 #include <openenclave/internal/time.h>
+#include <time.h>
 #include "../ocalls.h"
 
 static const uint64_t _SEC_TO_MSEC = 1000UL;
@@ -16,7 +17,8 @@ static uint64_t _time()
     if (clock_gettime(CLOCK_REALTIME, &ts) != 0)
         return 0;
 
-    return (ts.tv_sec * _SEC_TO_MSEC) + (ts.tv_nsec / _MSEC_TO_NSEC);
+    return ((uint64_t)ts.tv_sec * _SEC_TO_MSEC) +
+           ((uint64_t)ts.tv_nsec / _MSEC_TO_NSEC);
 }
 
 static void _sleep(uint64_t milliseconds)
